Moving steps into smaller classes
